requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
einops==0.6.0
ffmpeg-python==0.2.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
future==0.18.3
gradio==3.16.0
huggingface-hub==0.11.1
idna==3.4
kiwisolver==1.4.4
lightning-utilities==0.5.0
lpips==0.1.4 
matplotlib==3.6.2
multidict==6.0.4
numpy==1.24.1
omegaconf==2.3.0
open-clip-torch==2.9.1
opencv-python==4.7.0.68
packaging==22.0
Pillow==9.4.0
protobuf==3.20.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytorch-lightning==1.8.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
sentencepiece==0.1.97
six==1.16.0
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
yarl==1.8.2
